requirements.txt
asgiref==3.9.2
attrs==25.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==11.3.0
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.3
referencing==0.36.2
rest-framework-simplejwt==0.0.2
rpds-py==0.27.1
sqlparse==0.5.3
uritemplate==4.2.0
